Refactored MCPServerToolsForm
